\name{R2BayesX-package}
\alias{R2BayesX-package}
\alias{R2BayesX}
\alias{r2bayesx}
\docType{package}
 
\title{
  Estimate STAR Models with BayesX
}

\description{
  This package interfaces BayesX \dots
}

\details{
  \tabular{ll}{
    Package: \tab R2BayesX\cr
    Type: \tab Package\cr
    Version: \tab 0.1-1\cr
    Date: \tab 2011-05-30\cr
    License: \tab GPL-2 | GPL-3\cr
  }
}

\author{
  Thomas Kneib, Stefan Lang, Nikolaus Umlauf, Achim Zeileis.
}

\keyword{ package }

\examples{
  ## to see the package demos
  demo(package = "R2BayesX") 
}

\keyword{regression} 
